// web/src/api/generated/models/HeartEmotion.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Emotions defined by the M2 domain contract.
 * 
 * This is a closed value set but not metadata: the value lives in the
 * ProtectedPayload rather than a dedicated column. See M2-D06.
 * @export
 */
export const HeartEmotion = {
    LOVED: 'LOVED',
    SEEN: 'SEEN',
    APPRECIATED: 'APPRECIATED',
    SUPPORTED: 'SUPPORTED',
    GRATEFUL: 'GRATEFUL',
    HAPPY: 'HAPPY'
} as const;
export type HeartEmotion = typeof HeartEmotion[keyof typeof HeartEmotion];


export function instanceOfHeartEmotion(value: any): boolean {
    for (const key in HeartEmotion) {
        if (Object.prototype.hasOwnProperty.call(HeartEmotion, key)) {
            if (HeartEmotion[key as keyof typeof HeartEmotion] === value) {
                return true;
            }
        }
    }
    return false;
}

export function HeartEmotionFromJSON(json: any): HeartEmotion {
    return HeartEmotionFromJSONTyped(json, false);
}

export function HeartEmotionFromJSONTyped(json: any, ignoreDiscriminator: boolean): HeartEmotion {
    return json as HeartEmotion;
}

export function HeartEmotionToJSON(value?: HeartEmotion | null): any {
    return value as any;
}

export function HeartEmotionToJSONTyped(value: any, ignoreDiscriminator: boolean): HeartEmotion {
    return value as HeartEmotion;
}